fix: Se actuliza la funcionalidad crear/editar equipos
