--- public/pages/CrearEquipo.html
@@ -1,29 +1,0 @@
-<!DOCTYPE html>
-<html lang="es">
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>Crear Equipo</title>
-  <link rel="stylesheet" href="../styles/style.css">
-  <link rel="stylesheet" href="../styles/components/navbar.css">
-  <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@4.0.0/dist/css/bootstrap.min.css">
-  <link rel="stylesheet" href="../styles/components/button.css">
-</head>
-<body>
-    <nav id="navbar-placeholder"> </nav>
-    
-  
-    <form id="form-crear-equipo">
-      <div class="form-group">
-        <label for="equipoNombre">Nombre del equipo</label>
-        <input type="text" class="form-control" id="equipoNombre" placeholder="Ingrese el nombre del equipo" required>
-      </div>
-      <div id="personajes-disponibles" class="d-flex flex-wrap"></div>
-      <div id="button-placeholder"></div>
-      <button type="submit" class="btn btn-success mt-3"> Crear Equipo</button>
-    </form>
-  
-    <script type="module" src="/js/CrearEquipo.js"></script>
-  </body>
-  
-</html>
