+++ public/pages/EquipoForm.html
@@ -3,26 +3,25 @@
 <head>
   <meta charset="UTF-8">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>Editar Equipo</title>
+  <title>Formulario de Equipo</title>
   <link rel="stylesheet" href="../styles/style.css">
   <link rel="stylesheet" href="../styles/components/navbar.css">
   <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@4.0.0/dist/css/bootstrap.min.css">
   <link rel="stylesheet" href="../styles/components/button.css">
 </head>
-
 <body>
   <nav id="navbar-placeholder"></nav>
   <div class="container">
-    <form id="form-editar-equipo">
+    <form id="form-equipo">
       <div class="form-group">
         <label for="equipoNombre">Nombre del equipo</label>
-        <input type="text" class="form-control" id="equipoNombre" required>
+        <input type="text" class="form-control" id="equipoNombre" placeholder="Ingrese el nombre del equipo" required>
       </div>
       <div class="form-group">
         <label>Personajes del equipo (elige hasta 5)</label>
         <div id="personajes-equipo" class="d-flex flex-wrap"></div>
       </div>
-      <button type="submit" class="btn btn-primary">Guardar Cambios</button>
+      <button type="submit" class="btn btn-primary" id="btn-guardar-equipo">Guardar</button>
       <div id="button-placeholder"></div>
     </form>
 
@@ -31,6 +30,6 @@
     <div id="personajes-disponibles" class="d-flex flex-wrap"></div>
   </div>
 
-  <script type="module" src="/js/EditarEquipo.js"></script>
+  <script type="module" src="/js/CrearEditarEquipo.js"></script>
 </body>
 </html>
